0606:projectgotit 패키지 이름 변경
