+++ app/src/androidTest/java/com/example/projectgotit/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.prac7;
+package com.example.projectgotit;
 
 import android.content.Context;
