+++ app/src/main/java/com/example/projectgotit/BoardAdapter.java
@@ -1,55 +1,65 @@
-package com.example.prac7;
+package com.example.projectgotit;
 
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.graphics.ColorMatrix;
 import android.graphics.ColorMatrixColorFilter;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.CheckBox;
 import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
-
 import com.bumptech.glide.Glide;
+import com.example.projectgotit.network.RetrofitHelper;
+import com.example.projectgotit.network.RetrofitService;
 
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.HashSet;
 
-public class ScrapAdapter extends RecyclerView.Adapter<ScrapAdapter.ViewHolder> {
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
+import retrofit2.Retrofit;
+
+public class BoardAdapter extends RecyclerView.Adapter<BoardAdapter.ViewHolder> {
     private Context context;
     private List<BoardItem> itemList;
+    private OnItemClickListener listener;
 
-    public ScrapAdapter(Context context, List<BoardItem> itemList) {
+    public interface OnItemClickListener {
+        void onItemClick(int boardId);
+    }
+
+    public void setOnItemClickListener(OnItemClickListener listener) {
+        this.listener = listener;
+    }
+
+    public BoardAdapter(Context context, List<BoardItem> itemList) {
         this.context = context;
         this.itemList = itemList;
     }
 
     @Override
     public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.list_item, parent, false);
-        ViewHolder viewHolder = new ViewHolder(view);
-
-        return viewHolder;
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.list_item, parent, false);
+        return new ViewHolder(view);
     }
 
     @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
         BoardItem item = itemList.get(position);
-//        holder.imageView.setImageResource(item.getImageUrl());
         Glide.with(context).load(item.getImageUrl()).into(holder.imageView);
 
-        if (item.isFinished()) { // 종료된 소분이면
-            // 이미지 뷰를 흑백으로 만들기
-//            holder.imageView.setColorFilter(ContextCompat.getColor(holder.itemView.getContext(), android.R.color.darker_gray), android.graphics.PorterDuff.Mode.MULTIPLY);
-            // 이미지를 어둡게 설정
+        if (item.isFinished()) {
             ColorMatrix matrix = new ColorMatrix();
-            matrix.setSaturation(0); // 흑백으로 변환
+            matrix.setSaturation(0);
             ColorMatrix darkenMatrix = new ColorMatrix();
             darkenMatrix.set(new float[]{
                     0.5f, 0, 0, 0, 0,
@@ -59,15 +69,10 @@ public class ScrapAdapter extends RecyclerView.Adapter<ScrapAdapter.ViewHolder> 
             });
             matrix.postConcat(darkenMatrix);
             holder.imageView.setColorFilter(new ColorMatrixColorFilter(matrix));
-
-            // "종료" 텍스트 표시
             holder.textViewIsFinished.setVisibility(View.VISIBLE);
             holder.textViewIsFinished.setText("종료");
         } else {
-            // 조건에 맞지 않는 경우 이미지 뷰를 다시 원래 색상으로 변경
             holder.imageView.clearColorFilter();
-
-            // "종료" 텍스트 숨기기
             holder.textViewIsFinished.setVisibility(View.GONE);
         }
 
@@ -77,15 +82,56 @@ public class ScrapAdapter extends RecyclerView.Adapter<ScrapAdapter.ViewHolder> 
         holder.textViewMittingLocationText.setText(item.getLocation());
         holder.textViewParticipants.setText(item.getParticipants());
         holder.btnBookmark.setChecked(item.isBookmarked());
+
         holder.btnBookmark.setOnCheckedChangeListener((buttonView, isChecked) -> {
             item.setBookmarked(isChecked);
-            if (isChecked) {
+
+            SharedPreferences sharedPreferences = context.getSharedPreferences("loginPrefs", Context.MODE_PRIVATE);
+            String uid = String.valueOf(sharedPreferences.getInt("uid", -1));
+
+            // API를 호출할 객체 생성
+            Retrofit retrofit = RetrofitHelper.getRetrofitInstance();
+            RetrofitService retrofitService = retrofit.create(RetrofitService.class);
+            Call<Void> call = retrofitService.postScrap(uid, Integer.parseInt(item.getBid().split("\\.")[0]));
+
+            // API 호출
+            call.enqueue(new Callback<Void>() {
+                @Override
+                public void onResponse(Call<Void> call, Response<Void> response) {
+                    if (response.isSuccessful()) { // 200(정상적인 응답)이 왔을 때는 response.isSuccessful() 가 true
+                        Toast.makeText(context, "스크랩 되었습니다.", Toast.LENGTH_SHORT).show();
+                    }
+                }
+
+                @Override
+                public void onFailure(Call<Void> call, Throwable t) {
+                    Log.e("BoardAdapter", "Network Error: " + t.getMessage());
+                }
+            });
+            /*if (isChecked) {
                 addScrap(item);
             } else {
                 removeScrap(item);
-            }
+            }*/
         });
 
+        holder.itemView.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (listener != null) {
+                    String bidString = item.getBid();
+                    int bid;
+                    try {
+                        // 소수점을 기준으로 문자열을 잘라서 정수로 변환
+                        bid = Integer.parseInt(bidString.split("\\.")[0]);
+                        listener.onItemClick(bid);
+                    } catch (NumberFormatException e) {
+                        Log.e("BoardAdapter", "NumberFormatException: " + e.getMessage());
+                        // 적절한 오류 처리를 여기에 추가 (예: 사용자에게 오류 메시지를 표시)
+                    }
+                }
+            }
+        });
     }
 
     @Override
@@ -111,7 +157,7 @@ public class ScrapAdapter extends RecyclerView.Adapter<ScrapAdapter.ViewHolder> 
         editor.apply();
     }
 
-    public static class ViewHolder extends RecyclerView.ViewHolder {
+    public class ViewHolder extends RecyclerView.ViewHolder {
         private ImageView imageView;
         private TextView textViewTitle;
         private TextView textViewEachPrice;
@@ -131,25 +177,6 @@ public class ScrapAdapter extends RecyclerView.Adapter<ScrapAdapter.ViewHolder> 
             textViewMittingLocationText = itemView.findViewById(R.id.textViewMittingLocationText);
             btnBookmark = itemView.findViewById(R.id.btnScrap);
             textViewIsFinished = itemView.findViewById(R.id.textViewIsFinished);
-        }
-        public ImageView getImageView() { return imageView; }
-        public TextView getTextViewTitle() {
-            return textViewTitle;
-        }
-        public TextView getTextViewEachPrice() {
-            return textViewEachPrice;
-        }
-        public TextView getTextViewCategory() {
-            return textViewCategory;
-        }
-        public TextView getTextViewParticipants() {
-            return textViewParticipants;
-        }
-        public TextView getTextViewMittingLocationText() {
-            return textViewMittingLocationText;
-        }
-        public CheckBox getBtnBookmark() {
-            return btnBookmark;
         }
     }
 }
