+++ app/src/main/java/com/example/projectgotit/MyPageActivity.java
@@ -1,4 +1,4 @@
-package com.example.prac7;
+package com.example.projectgotit;
 
 import android.content.Context;
 import android.content.Intent;
@@ -7,15 +7,14 @@ import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
-import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.prac7.network.RetrofitHelper;
-import com.example.prac7.network.RetrofitService;
+import com.example.projectgotit.network.RetrofitHelper;
+import com.example.projectgotit.network.RetrofitService;
 
 import java.util.HashMap;
