+++ app/src/main/java/com/example/projectgotit/PointHistoryItem.java
@@ -1,4 +1,4 @@
-package com.example.prac7;
+package com.example.projectgotit;
 
 public class PointHistoryItem {
     private String changeReason;
