+++ app/src/main/java/com/example/projectgotit/ScrapActivity.java
@@ -1,25 +1,16 @@
-package com.example.prac7;
+package com.example.projectgotit;
 
 import android.content.Context;
-import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.EditText;
-import android.widget.ImageButton;
-import android.widget.ImageView;
-import android.widget.Spinner;
-import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.prac7.network.RetrofitHelper;
-import com.example.prac7.network.RetrofitService;
+import com.example.projectgotit.network.RetrofitHelper;
+import com.example.projectgotit.network.RetrofitService;
 
 import java.util.ArrayList;
 import java.util.HashMap;
