+++ app/src/main/java/com/example/projectgotit/network/RetrofitHelper.java
@@ -1,4 +1,4 @@
-package com.example.prac7.network;
+package com.example.projectgotit.network;
 
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
