+++ app/src/test/java/com/example/projectgotit/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.prac7;
+package com.example.projectgotit;
 
 import org.junit.Test;
